feat: Solucionado el tema del salto. Se consiguen cargar los datos pero no bien del todo. Falta poder guardar la salud.
