+++ Unity/BrokenFaith/Assets/Miquel/Scripts/Partida.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9f03c1eea001dae44b384e98c0749a72
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
