+++ Unity/BrokenFaith/Assets/Miquel/Scripts/Partida/Partida.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9ab0f1f9b340414496efb83bb02fdde
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
