+++ Unity/BrokenFaith/Assets/Soraki/Scripts/Guardar Datos/Progreso.cs
@@ -47,9 +47,9 @@ public class Progreso
         Debug.Log("Guardado en " + ruta);
 
     }
-    public ContenedorDeSalud CargarPartida()
+    public Partida CargarPartida()
     {
-        string ruta = Path.Combine(Application.dataPath, "Guardado.json");
+        /*string ruta = Path.Combine(Application.dataPath, "Guardado.json");
         if (File.Exists(ruta))
         {
             string datosGuardados = File.ReadAllText(ruta);
@@ -57,7 +57,14 @@ public class Progreso
             string datosJson = JsonUtility.ToJson(progresoCargado);
             return progresoCargado;
         }
-        return null;
+        return null;*/
+
+        string ruta = Path.Combine(Application.dataPath, "Guardado.xml");
+        XmlSerializer serializer = new XmlSerializer(typeof(Partida));
+        using (StreamReader reader = new StreamReader(ruta))
+        {
+            return (Partida)serializer.Deserialize(reader);
+        }
     }
     public void ObtenerDatos(List<Salud> pj)
     {
@@ -85,53 +92,4 @@ public class Progreso
 public class ContenedorDeSalud
 {
     public EstadoJugador Salud;
-}
-
-// Añadidas las dos funciones por Miquel Grau el 24/02/24 para poder serialiizar el inventario
-public class Partida
-{
-    /*[XmlArray("Salud")]
-    [XmlArrayItem("Brazos", typeof(Brazos))]
-    [XmlArrayItem("Cabeza", typeof(Cabeza))]
-    [XmlArrayItem("Cuerpo", typeof(Cuerpo))]
-    [XmlArrayItem("Piernas", typeof(Piernas))]
-    [XmlArrayItem("Torso", typeof(Torso))]
-    public List<Salud> Salud { get; set; }*/
-
-    [XmlArray("Inventario")]
-    [XmlArrayItem("Clave", typeof(Clave))]
-    [XmlArrayItem("Equipo", typeof(Equipo))]
-    [XmlArrayItem("Item", typeof(Item))]
-    [XmlArrayItem("Material", typeof(Material))]
-    [XmlArrayItem("Medicina", typeof(Medicina))]
-    [XmlArrayItem("Municion", typeof(Municion))]
-    [XmlArrayItem("Pieza", typeof(Pieza))]
-    [XmlArrayItem("Pista", typeof(Pista))]
-    public List<Item> Inventario { get; set; }
-
-    [XmlElement("Position")]
-    public Vector3Serializable Position { get; set; }
-
-    public Partida()
-    {
-        //Salud = new List<Salud>();
-        Inventario = new List<Item>();
-    }
-}
-
-public class Vector3Serializable
-{
-    public float X { get; set; }
-    public float Y { get; set; }
-    public float Z { get; set; }
-
-    public Vector3Serializable() { }
-
-    // Constructor para convertir de Vector3 a Vector3Serializable, si es necesario
-    public Vector3Serializable(Vector3 vector)
-    {
-        X = vector.x;
-        Y = vector.y;
-        Z = vector.z;
-    }
 }
