+++ Unity/BrokenFaith/ProjectSettings/EditorBuildSettings.asset
@@ -12,5 +12,7 @@ EditorBuildSettings:
     path: Assets/Victor/Scenes/Blocking+ Mapa1.unity
     guid: 48509604997407740b7d843d9ca47994
   m_configObjects:
+    Unity.XR.Oculus.Settings: {fileID: 11400000, guid: 384319fd4ab37284f83f572f23b23db6,
+      type: 2}
     com.unity.xr.openxr.settings4: {fileID: 11400000, guid: 704fab20f39788a439fbe4a27bd83407,
       type: 2}
